Started fiddling with the html a bit for apps.
I'm realizing I need to take the approach of make the HTML and then make
the css to format it. that seems to be the way the systems are designed to
work. So I've started trying to add some content that I will then learn
how to form how I want. I also made static content to start with. I'll
learn how to do the fancy stuff with dynamic content as soon as I learn
how to even get a page up.