Created post objects.
I've added a little more to this. Now there are "posts" that will
encapsulate all the information for a post. I've run into a problem
however which I'm not sure what is causing it. That sentence got
weird. I think the user not being able to access the git repo is
basically the problem. Which is probably do to clever tricks with
the NFS mount not being quite as clever as I want.