Apps now are devlogs! Wait.
It is mostly done as a matter of crossed wires. But I've added the
feature where commits are posted as blog entries to the apps page.
Not entirely intentional. But I was already down that rabbit hole
so I kept going. I'll make the appropriate changes soon to move the
git commit auto blogging over to the dev log page, which I've yet
to make. ^^;

But it's progress! The site is starting to have some idea of the
content that I want in it available. And I'm slowly starting to
understand how to piece it all together. It is my first flask app
after all. Man, never thought I'd be making web apps. I suppose I
never really fancied myself as a mobile developer either. But here
we are. Making kivy and flask apps.

TO THE KING!