Works now!
I got the posts.py file working like I want it to now. Turns out
there was a trailing '\n' causing all the problems. But I decided
to upgrade the repos file to use json so I could add more info to
the projects.

In looking at this, I might release this as it's own project in the
future. Autoblogging via git commit messages is pretty knifty and
I think more people than just me would enjoy it. We'll see. Right
now I'm jsut happy to get all the projects in to the system.