APP_ROOT
A small fix, but I have updated the pathing for resources. While
resource/projects.json worked fine for the dev enviorment, it was
failing in production. So I added a bit that I found on stack
overflow: https://stackoverflow.com/q/14825787

This introduced me to getting the location of the current file and
saving it as APP_ROOT. This should work in both dev and prod.