Apps now does it's own thing.
I've decided to leverage the README.md file for the summary of the
app. I'll probably end up doing something more interesting in the
future, but this is certainly a functional first step, something
that allows me to get this website up sooner than later.

I've also re-arranged things a bit. I've moved my git-post/posts.py
file into utility-scripts/ folder. It and copy-markdown.py are the
occupants of my new utility-scripts folder. These are scripts I'll
have to run periodically for now to get the data to show.