Style is coming!
I've got style actually starting to appear on the website now. I
banged my head against the flask-bootstrap templates for a while
but in the end decided to not use the base templates and instead
create my own. This requires a little more effort, but it allows me
a lot more control. Since HTML is still something I'm trying to
master, keeping it simple is a great help.

I've downloaded a bootstrap css style. I can easily change the
style of my whole site but just swapping these files out. Pretty
neat really! Right now I've just choosen a simple black style. I'm
found of sites that are primarily dark colors, it feels less
straining to look at. As I add pictures and things, it may be less
of a deal and I'll probably switch to a lighter style, but this
works for now.