Things are looking good!
I've made a huge pile of changes with this commit. Pictures have
been added to most of the apps on the apps page. the pictures will
naturallly flow behind the text for each application. I put down
some basic pieces for about and contact, though some tweeking is
still necessary. I've added a discription for each app.

Over all this is really starting to close in on being a fleshed out
website the way have been envisioning/desiring. I'm pleased! Just
a little more and I'll be able to make it public facing.