Temporary house viewer.
I've added a few things and changed the default home page to allow
for family to see the pictures of houses we are looking at. This
will probably go away in the near future, but it is convenient for
now.

It took a little effort, but I even got lazy loading working on
this. So the images now load when you call for them, and no
earlier. That should keep it's load time low and also keep it very
responsive.