Presentable.
I think I've got it to a basic presentable level. I set all but
the projects tab to fill only 50% of the screen so it's not spread
so thin. Projects will probably need some formating of their own,
but for now this will do. I also changed the name from "apps" to
"projects" since that more accurately reflects the nature of them.
It isn't so much individual programs as much as it is various
projects that I am working on.