NCS Player and blog.
I decided to add a blog. More details on why in my blog. Basically
git posts and the project pages weren't quite sufficent for what I
needed. I had content I wanted to post but no where to put it since
all content had to be directly related to one of my projects.

Also new in this commit is the NCS Player. NCS Player is a proof of
concept project. I have plans to make a video streaming app for
independant content creators, but first I wanted to experiment
around with something a little more constrained in scope.