Backup and refresh
I added two new utility scripts to keep everything up to date and
safe. The first is a script that goes through all of the repos
locally on carbon and pulls down the latest versions from mycroft.
After doing this then the website's info can be updated, but that
will be part of a pipeline sort of thing I develop later. The
important thing is that I don't have to do each independantly now.

The second script makes an archive. It tarballs all the repos
together and sends a copy over to mycroft. In it's final form it
will also delete the tarball from carbon, but I had a hard drive
scare a bit ago and so I don't want to leave all my data on
mycroft. I plan on running this about once a week. Version control
is plenty good, but I just want a backup my backup as the saying
goes. Eventually this will upload to the cloud somewhere.