Json blog refinement.
Just a few little tweaks to get the json blog stuff working fully
and correctly. I landed on using a hybrid system of json files and
mark down files. Becuase writing a markdown file inside a json file
is a terrible idea.  So instead, it gets both. You can specify in
the json file what the markdown file is for the blog post.