Snow removal!
I added a system for creating snow shoveling invoices for my mom.
She has me shovel snow outside the sorority she manages on occation
and I wanted a nice simple way to keep track and make invoices for
billing purposes. These are the invoices that get sent to the board
so they need to look some what professional. I've also set it up to
be easy to add each clearing and then re-generate the pdf. Now all
I have to do is edit snow.csv, adding the days I shoveled snow and
then at the end of the month, vist the PDF and down load it.
Because of the rarity in which I need this feature, I'll probably
not push it to carbon.