Resume is now available, but needs some love.
I've added a resume feature to my website. Like everything it is
data driven. I have a template for the skeleton of the resume but
a JSON file is loaded that populates it with data. This commit is
just a basic setup. I've copied in all of my information from my
previous resume into the json file and formated it in a similar
way. Honestly it looks somewhere between crap and bland. But it's a
first step. Now that I've got the skeleton and data there I will
clean up the template to make it look nicer and update the info
in resume.json to be more accurate of recent developments.