Added some basic home automation.
I wanted to be able to control my kitchen lights from IFTTT, so
I wrote a script that will allow me to pass through the commands
with out actually opening my home network to arbitrary html
requests. That would be bad.

This isn't entirely great since theoretically anyone could find
this and take control of our kitchen lights. I'm not to concerned
because that's a worthless target. They only draw 1A 12V so it's
not like it can be used like our AC for malicious power grid
attacks or anything like that. The worst they could do is make
them constantly flash.