Updates to blog linking.
There was a problem that was outstanding for a while that I've been
wanting to fix but have never gotten around to it. blog posts have
tags that when you click on them link you to a tags page so you can
see all other posts with those tags. That last piece of it wasn't
working. It would take you to a blank page. Got that sorted. I also
moved the blog subsystem to its own blue print. This drops my
__init__.py file down to about 100 lines of code. Probably about
half the size it was when I started this endevour. Jolly good.