Python3 and Tracking.
I've decided to switch to python3. This actually only has effects
on one point of the code base, StringIO has been replaced with just
io. Considering that is for the "snow" blueprint I may not even
keep that around so the end result is a python agnostic code base.

I've also added basic tracking. Now every request gets recorded
into a database that I can query later. I don't know if anyone sees
my site but this should give me some idea and using geo-ip I can
potentially get some geographical data out of it. Since the purpose
of this website is to be a portfolio it would be nice to see who
is looking at it from where. The who will be quite challenging, but
the IP is enough for the where. Once I have that information then
I'd like to make some sort of display.