Trying to rebuild the site in react.
I've made a fair bit of progress for a single afternoon working in
react. I'll give it a little more effort tomorrow evening. Right
now I need to setup the ajax part of the site so that it can lazy
load the blog posts and such. But it is turning into a nice single
page application. Fancy!