It's all coming together now.
I've got the basics of the Developer's log working. It will display
projects and you can click on the projects to see the git commits,
such as this one, for that project. The great thing is that the
rest of the site is mostly a copy of this. Project and blog posts
are basically the same thing as the dev log so I just need to make
those components in their appropriate stylings.

React isn't to hard all in all. I do greatly appreciate having web
components in my site at last. It's been over a year since I first
learned about them and I've done vue and I've done polymer but
never got around to doing them in my own site. If it weren't for
the job offer at planning center I probably would have just done
vue. I'm glad I got to try react, I think it's cleaner, at least
with JSX. I wonder if there is a JSX Flask pluggin, that would be
pretty cool. Well, onward!