Blog is joggin.
I've got the web log portion of the site up and running now. It
took an extra little tick because I render the markdown into HTML
on the server. I could do this in react but I didn't want to load
any modules in order to get my code to work. For now I'm using pure
react while I'm just learning my way around. I'm not sure though in
the end which I'd prefer, react parsing the mark down or python
parsing the mark down. But the important thing is that it works for
what I want right now.