Projects is up and working.
I think I've got all the functional aspects except for Resume up
and working again now. There are a few formating things that could
be tightened up but this is a good point to commit. I also need to
update all the data that backs the website. More recent gits more
recent readmes. Of course those are generated so it's not a lot of
effort but still it is something that I need to do.