Merging of Home and About
I've merged these two pages as they were both rather sparce and a
little redundant. So I've merged them into one page. The other
reason is to get the about me front and center when some one comes
to the page to begin with. Make it a little more me focused I
suppose you could say. But hopefully in the "You're here to hear
about me, here's my info" way and not the braggart way.