Updates to post generation.
I made a few tweaks to the script that turns my commit messages
into devlog posts. I set it up to only harvest my commits so I am
not hosting other people's commits as if they were mine.

Also, it would seem the script does not want to run on my website
while it is in a different un-commited branch. So hopefully it will
work correctly as soon as I merge this into master.