I have some bits working.
It is a radically different way of doing things than what I have
been doing but I think I'll get the hang of it pretty quickly. I've
got the parts that I couldn't figure out previously working, though
the site is just a couple of p tags right now. I'm going to
statically host the json files for now. It is simple enough to just
update them from the OS when needed. That allows me to focus on
getting the site up for now.