Blog is now in.
I've got the basics for the blog now in. It will take a little more
effort to include the tags and cover arts. But that is true of all
of the tabs: More work needed. The important thing is react router
is now catching everything and navigating you to the pags. In fact
only the resume page and the blog tags pages are yet to be made.
Great progress! Hopefully next time I'm working on this all have
the skeleton fully fleshed out and I will be able to start putting
some flesh on to it.