The flesh is there.
or at least the content is. All the parts are present. The posts,
the tags that connect the posts, the resume, etc. I've also gotten
the home page filled with content and the banner image in place.
All that is left to do now is to get all the styling that is needed
in place so it doesn't look really bad. The week after that I'll
look into getting it hosted on nginx on midgard. And then I'll try
my hand at some ruby on rails. That should be interesting.