Making good progress on styling.
I'm making a project link component since I will be using it from
both the projects tab and the dev log tab. Right now I'm to the
point where I need to bring in the pictures. So I'm going to save
here and then go fetch them and come back.