Python API almost up to new standard.
I've pasted what I originally implemented in Ruby and now have more
functionality in the form of hosting files for projects and posts.
The last thing I want to do is create a feed of some sort. Probably
both rss and a json object feed that aggrigates across logs and
posts to show what activity is happening on my website, and
therefore what I am doing.