Rewriting to use API
Now that I have the python API fleshed out I'm rewiring the app to use
the API to get its information. Mostly just changing urls, but here and
there a few more indepth tweaks. Though Doing this has also pointed out
to me one of the things I missed: Tags. Blog tags. I also will need to
extend some of the React components to take advantage of new features.
Most notably the files attached to blog posts. This is a feature I had
to add to be able to share the work I've been doing on Simple Systems.
Its been over a month I think since I finished that series and none of
it is on my website yet. But The reason for that is I didn't have a good
mechanism for sharing the code files. This primarily because I wanted it
to be automatically wired in and not something I would have to manually
enter for each post.

Anyway. It is bed time, but a lot of the website has been updated.
Tomorrow I'll try to finish it and then bring the Ruby API up to snuff
and push both out to midgard.