Python API now has Feeds!
I have added an RSS2.0 compliant feed and a feed json object to the api
to allow for feed readers and so I can put a feed up on the home page to
display a quick look at what I am doing.

This has made me also realize that I should have in page anchors on my
dev logs, so that one can link to a specific commit. The items in the
feed right now just link to the project's page with all of the commits.