I've found a problem while making the unit tests.
I'm not consistent in the way I form the URLs in the API. I have made
the test consistent for logs, I need to do it for projects yet, and then
I need to fix the API so it provides consistent URLs as well.

But the good news is I'm about half way through the API validator.
Should be done tomorrow morning depending on how long it takes me to
make the muffins.