I said I wanted to learn to think in ruby right?
Man, I didn't think it would require such a different mind sets. And
while it is true that the logic stays the same Ruby is a much more
"magical" language, which admittedly I do not like. I don't like it when
Java tries to be "magical" I don't like it when Python tries to be
"magical" as being "magical" usually means hiding details so it is
impossible to follow the flow of information.

In this case the problem I'm trying to wrap my head around is implicity
returns. The last statement of a ruby method is automatically returned.
Interesting approach but they also seem to treat normal returns
differently from python and other C descendants. I'm going to have to
study this a little further to understand what is really going on here
and how I get my method to return when I want it to.