Ruby Logs now compliant.
I gotten ruby logs to now be compliant with the API. Found some un-used
code in they python api while I was at it. This just leaves the rss and
json feeds left to implement and then the Ruby API is fully in
compliance with the Python API. I'll be happy when I have finished this
and can actually get it out to my website at long last.