Got it!
Both the Ruby API and the Python API fully pass the unit tests. The two
are identical as far as I can tell. Now then, time to do some more
updates to the front end. Specifically I want to have a feed display
what I've been up to on the home page.