API Selector exists!
I'm going to have to figure out how to do the nginx switching based on
cookies, but the basics are here. I've got a little fixed box that you
can use to select whether you want to use the Python API or the Ruby
API. Which reminds me, I need to stand up the python api service.