SSGET
ssget a little wrapper method I wrote to insert which API to make a
request to. So I'm still using cookies like I had planned, but I
realized that the API request was going to api.studiosleepygiraffe.com
and not to studiosleepygiraffe.com. Ergo, the cookie would not make its
way to nginx for routing. So I'm going to have to do url based routing,
which allows for explicit routing anyway which is nice.

I've also added a python api service file.